+++ src/bridge/index.ts
@@ -5,6 +5,7 @@
  */
 export * from "./asset-recipient.js";
 export * from "./client.js";
+export * from "./deposit-status.js";
 export * from "./extradata.js";
 export * from "./l1-builder.js";
 export * from "./operation-id.js";